pyproject.toml
# Responsibility: Define Python package metadata, dependencies, and tooling entry points for kibo-converter.

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "kibo-converter"
version = "0.1.0"
description = "Batch image conversion with job definitions, progress, and logging."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.6",
    "Pillow>=10.0",
    "pillow-heif>=0.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
    "pytest-mock>=3.12",
]

[project.scripts]
kibo-converter = "kibo_converter.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
qt_api = "pyqt6"
